requirements.txt
python-dotenv==1.1.0
requests==2.32.3
beautifulsoup4==4.13.4
openai==1.12.0
moviepy==1.0.3
pillow==11.2.1
pydub==0.25.1
elevenlabs==1.57.0
